# src/main/resources/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Tiny Cerberus Bot",
        "description": "Discipline Telegram Chat bot",
        "version": "v1.2.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:18080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/admin/bot/statistics/chats/{chatId}": {
            "get": {
                "tags": [
                    "statistics-controller"
                ],
                "operationId": "getChatStatistics",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatMemberStatisticsView"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/bot/statistics/chats/{chatId}/warns": {
            "get": {
                "tags": [
                    "statistics-controller"
                ],
                "operationId": "getChatWarnStatistics",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatWarnStatisticsView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/bot/statistics/chats/{chatId}/top": {
            "get": {
                "tags": [
                    "statistics-controller"
                ],
                "operationId": "getChatTop",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopChatMemberView"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/bot/statistics/chats/{chatId}/silent-members": {
            "get": {
                "tags": [
                    "statistics-controller"
                ],
                "operationId": "getSilentMembers",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SilentChatMemberView"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/bot/statistics/chats/{chatId}/all": {
            "get": {
                "tags": [
                    "statistics-controller"
                ],
                "operationId": "getCompleteChatStatistics",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatMemberStatisticsView"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChatMemberStatisticsView": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "messageCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ChatUserWarnStatisticsView": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "telegramId": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "warnCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ChatWarnStatisticsView": {
                "type": "object",
                "properties": {
                    "chatId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "warnLimit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChatUserWarnStatisticsView"
                        }
                    }
                }
            },
            "TopChatMemberView": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "messageCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "SilentChatMemberView": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayName": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}